+++ packages/webapp/src/inductions/components/index.ts
@@ -1,3 +1,4 @@
 export * from "./get-an-invite-cta";
 export * from "./pending-inductions";
 export * from "./induction-journeys";
+export * from "./induction-journey-container";
