+++ packages/webapp/src/pages/induction/[id].tsx
@@ -1,4 +1,3 @@
-import React, { useMemo } from "react";
 import { useRouter } from "next/router";
 
 import {
@@ -10,13 +9,9 @@ import {
 } from "_app";
 import {
     getInductionStatus,
-    InductionRole,
+    InductionJourneyContainer,
     InductionStatus,
-    InviteeJourney,
-    InviterWitnessJourney,
-    ThirdPartyJourney,
     useGetInductionWithEndorsements,
-    useInductionUserRole,
 } from "inductions";
 
 export const InductionDetailsPage = () => {
@@ -35,45 +30,13 @@ export const InductionDetailsPage = () => {
     const induction = data?.induction;
     const endorsements = data?.endorsements || [];
 
-    const userRole = useInductionUserRole(endorsements, induction);
     const status = getInductionStatus(induction, endorsements);
 
-    const renderInductionJourney = useMemo(() => {
-        if (!induction) return "";
-        switch (userRole) {
-            case InductionRole.Inviter:
-            case InductionRole.Endorser:
-                return (
-                    <InviterWitnessJourney
-                        endorsements={endorsements}
-                        induction={induction}
-                        inductionStatus={status}
-                    />
-                );
-            case InductionRole.Invitee:
-                return (
-                    <InviteeJourney
-                        endorsements={endorsements}
-                        induction={induction}
-                        inductionStatus={status}
-                    />
-                );
-            default:
-                return (
-                    <ThirdPartyJourney
-                        endorsements={endorsements}
-                        induction={induction}
-                        inductionStatus={status}
-                    />
-                );
-        }
-    }, [induction, endorsements, status, userRole]);
+    const notFound =
+        status === InductionStatus.Invalid ||
+        status === InductionStatus.Expired;
 
-    if (
-        !isLoading &&
-        (status === InductionStatus.Invalid ||
-            status === InductionStatus.Expired)
-    ) {
+    if (!isLoading && notFound) {
         return (
             <RawLayout title="Invite not found">
                 <CallToAction
@@ -95,7 +58,13 @@ export const InductionDetailsPage = () => {
             {isLoading ? (
                 <Card title="Loading...">...</Card>
             ) : (
-                renderInductionJourney
+                induction && (
+                    <InductionJourneyContainer
+                        induction={induction}
+                        endorsements={endorsements}
+                        status={status}
+                    />
+                )
             )}
         </SingleColLayout>
     );
